refactor: nutrition => stateful class component, foodCard => stateless functional component
